chore: move tokenBucket type to the algorithm file
